Removed Isolates
Isolates were removed which resulted in good results for ERGM. However, the GOF of geodesic distance seems to indicate an issue. 

Walktrap is the best algorithm for clustering (with and without isolates)

Walktrap directed is better than walktrap undirected. Why? 

Clusters can be significant also without isolates. Without isolates, it depends on which significance level we choose (p-value tends to be between 0.055 and 0.09)